+++ packages/stacks-svelte/src/components/UserCard/UserCard.svelte
@@ -8,7 +8,7 @@
     import Avatar, { type Size as AvatarSize } from "../Avatar/Avatar.svelte";
     import Bling from "../Bling/Bling.svelte";
     import Icon from "../Icon/Icon.svelte";
-    import { IconPerson } from "@stackoverflow/stacks-icons/icons";
+    import { IconPerson } from "@stackoverflow/stacks-icons-legacy/icons";
 
     interface Props {
         /**