+++ packages/stacks-classic/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @stackoverflow/stacks
 
+## 3.0.0-beta.2
+
+### Minor Changes
+
+- [#2015](https://github.com/StackExchange/Stacks/pull/2015) [`66bfd40`](https://github.com/StackExchange/Stacks/commit/66bfd409b4643db3970457eb884066aa799c15d9) Thanks [@dancormier](https://github.com/dancormier)! - Updated colors for SHINE
+
+- [#2028](https://github.com/StackExchange/Stacks/pull/2028) [`c9bd0f6`](https://github.com/StackExchange/Stacks/commit/c9bd0f6ed0286bd2b1b2935662f96c537445bf12) Thanks [@dancormier](https://github.com/dancormier)! - Update typography for new SHINE design system
+
+### Patch Changes
+
+- [#2037](https://github.com/StackExchange/Stacks/pull/2037) [`73c8edd`](https://github.com/StackExchange/Stacks/commit/73c8edd05eacee5d26e9267279a1cb5b32923710) Thanks [@dancormier](https://github.com/dancormier)! - Add legacy sm, lg border radius custom properties
+
+## 3.0.0-beta.1
+
+### Minor Changes
+
+- [#2026](https://github.com/StackExchange/Stacks/pull/2026) [`2a4a741`](https://github.com/StackExchange/Stacks/commit/2a4a74100534b2020aa7e9ebecd956a2aa15a74e) Thanks [@ttaylor-stack](https://github.com/ttaylor-stack)! - Update border radius for new SHINE design system
+
+    BREAKING CHANGE:
+    - bar-sm and bar-lg are deprecated use bar-md instead
+
+- [#2030](https://github.com/StackExchange/Stacks/pull/2030) [`da8616a`](https://github.com/StackExchange/Stacks/commit/da8616a618e0d0f36f768a7f6600b5b6e129b314) Thanks [@ttaylor-stack](https://github.com/ttaylor-stack)! - Update popover component to match SHINE styles
+
+    BREAKING CHANGES:
+    - `s-popover--arrow` css class has been removed
+
+### Patch Changes
+
+- [#2019](https://github.com/StackExchange/Stacks/pull/2019) [`e9b8885`](https://github.com/StackExchange/Stacks/commit/e9b888585e2d145e4216ea413904bba83a4faa80) Thanks [@giamir](https://github.com/giamir)! - **Avatar Updates:**
+    - **Stacks Classic**: Made avatar border squared. Introduced `.s-avatar--indicator` class for positioning an activity indicator element in relation to the avatar.
+    - **Stacks Svelte**: Introduced a new `status` prop. When set to `online` it will show an activity indicator on the avatar.
+
+- [#2027](https://github.com/StackExchange/Stacks/pull/2027) [`4bd5612`](https://github.com/StackExchange/Stacks/commit/4bd561230e43bb411a58f1cf975c0a5f62a205fe) Thanks [@giamir](https://github.com/giamir)! - **Activity Indicator Updates:**
+    - **Stacks Classic**: Updated activity indicator component to the new visual language (default 16x16). Introduced `.s-activity-indicator__sm` variant (10x10).
+    - **Stacks Svelte**: Added new `ActivityIndicator` component.
+
 ## 3.0.0-beta.0
 
 ### Major Changes